requirements.txt
python-telegram-bot==21.6
python-dotenv
yt-dlp
fastapi
uvicorn
